move files into rollup subdirectory
