+++ packages/firestore/rollup/remove-asserts.ts
