+++ packages/firestore/rollup/rename-internals.ts
